refactor: 궁성 영역 판단을 Position이 아닌 Country에서 수행하도록 수정
